# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nozay Badminton API",
    "version": "1.0.0",
    "description": "Verifiable contract between apps/api, apps/admin and apps/storefront"
  },
  "paths": {
...
    "/members": {
      "get": {
        "summary": "List and filter members",
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "page",
                        "limit",
                        "pages"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/expenses": {
      "post": {
        "summary": "Create a new expense note",
        "tags": [
          "Expenses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created expense details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateExpenseInput": {
        "type": "object",
        "required": [
          "seasonId",
          "description",
          "category",
          "amount",
          "emitterName"
        ],
        "properties": {
          "seasonId": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "amount": {
            "minimum": 1,
            "type": "integer"
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "emitterName": {
            "minLength": 1,
            "type": "string"
          },
          "memberId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    }
  }
}